Converted things to .NET Framework 4.8 and MonoGame 3.8. (#30)
A few things were tweaked to get this working. The EULA is currently broken. Updates also don't work, but they haven't been working for a while. The ability to submit a bug or feature request is also broken, though perhaps people should just come here to do that. (And maybe putting in a link to here is a good short- or medium-term fix.)

Co-authored-by: RB Whitaker <RB.Whitaker@asirobots.com>