Did a lot of cleanup for things that were easy to verify.
There is a ton of cleanup that needs to be done in this repository to bring it in alignment with modern C# conventions. The changes here were to do a few things that were simple and obvious, though I left a lot of other changes undone. It is a step in the right direction at least.